graph-designer-agent/pyproject.toml
[project]
name = "graph-designer-agent"
version = "0.1.0"
description = "ADK Agent for Graph Schema Design and Spanner Deployment"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-spanner>=3.40.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
